requirements.txt
Authlib==0.11
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
distlib==0.3.4
filelock==3.6.0
Flask==2.0.3
gdown==4.4.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Naked==0.1.31
platformdirs==2.5.2
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.14.1
pyparsing==3.0.8
PySocks==1.7.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
shellescape==3.8.1
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.0
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.1.2
zipp==3.8.0
